Moved TestUtil to another package.
